Removing tabmanager controller
