--- src/main/java/com/flipper/controllers/TabManagerController.java
@@ -1,59 +1,0 @@
-package com.flipper.controllers;
-
-import net.runelite.client.util.ImageUtil;
-import net.runelite.client.ui.NavigationButton;
-
-import com.flipper.helpers.UiUtilities;
-import com.flipper.views.TabManager;
-import com.flipper.views.buys.BuyPage;
-import com.flipper.views.flips.FlipPage;
-import com.flipper.views.login.LoginPage;
-import com.flipper.views.margins.MarginPage;
-import com.flipper.views.sells.SellPage;
-
-import net.runelite.client.ui.ClientToolbar;
-
-public class TabManagerController {
-    private TabManager tabManager;
-    private NavigationButton navButton;
-    private ClientToolbar clientToolbar;
-
-    public TabManagerController(ClientToolbar clientToolbar) {
-        this.clientToolbar = clientToolbar;
-        tabManager = new TabManager();
-        setUpNavigationButton();
-    }
-
-    private void setUpNavigationButton() {
-        navButton = NavigationButton
-            .builder()
-            .tooltip("Flipper")
-            .icon(ImageUtil.getResourceStreamFromClass(
-                getClass(), 
-                UiUtilities.flipperNavIcon  
-            ))
-            .priority(4)
-            .panel(tabManager).build();
-        clientToolbar.addNavigation(navButton);
-    }
-
-    public void changeToLoggedInView(
-        BuyPage buyPage, 
-        SellPage sellPage, 
-        FlipPage flipPage, 
-        MarginPage marginPage,
-        Runnable changeToLoggedOutView
-    ) {
-        tabManager.renderLoggedInView(
-            buyPage, 
-            sellPage, 
-            flipPage, 
-            marginPage,
-            changeToLoggedOutView
-        );
-    }
-
-    public void changeToLoggedOutView(LoginPage loginPage) {
-        tabManager.renderLoggedOutView(loginPage);
-    }
-}
